+++ CHANGELOG.md
@@ -11,6 +11,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 -   Set composer platform PHP version to 5.6 to ensure package compatibility (#5266)
 
+### Changed
+
+-   Multi-Form Goal Blocks now auto-focus the Progress Bar on insert (#5364)
+
 ## [2.9.0-alpha.2] - 2020-10-09
 
 ### Added